modules/network_scan.py
"""
Network Scan module for CYBERNADE.
Will be fully implemented in Phase 3.
"""
import logging
import socket
from core.database import db

class NetworkScanner:
    """Network scanner for CYBERNADE."""
    
    def __init__(self):
        """Initialize the network scanner."""
        logging.info("Network scanner initialized")
        
        # Flag to determine if nmap is available
        self.nmap_available = False
        
        try:
            # Try to import nmap
            import nmap
            self.nm = nmap.PortScanner()
            self.nmap_available = True
            logging.info("python-nmap imported successfully")
        except ImportError:
            logging.warning("python-nmap not available, functionality will be limited")
    
    def ping_sweep(self, network):
        """Discover active hosts in a network.
        Stub method for Phase 3.
        
        Args:
            network (str): Network to scan (e.g., 192.168.1.0/24)
            
        Returns:
            list: List of active hosts
        """
        # This is a stub that will be implemented in Phase 3
        logging.info(f"Ping sweep requested for {network}")
        
        # Save to database
        db.log_activity("network", "ping_sweep", {"network": network})
        
        # Return placeholder results
        return {
            "network": network,
            "status": "Ping sweep functionality will be implemented in Phase 3"
        }
    
    def port_scan(self, target, ports="1-1000", scan_type="SYN"):
        """Scan ports on a target.
        Stub method for Phase 3.
        
        Args:
            target (str): Target to scan
            ports (str): Ports to scan (e.g., "1-1000", "22,80,443")
            scan_type (str): Type of scan (SYN, TCP, UDP)
            
        Returns:
            dict: Results of the scan
        """
        # This is a stub that will be implemented in Phase 3
        logging.info(f"Port scan requested for {target} (ports: {ports}, type: {scan_type})")
        
        # Save to database
        db.log_activity("network", "port_scan", {
            "target": target,
            "ports": ports,
            "scan_type": scan_type
        })
        
        # Return placeholder results
        return {
            "target": target,
            "ports": ports,
            "scan_type": scan_type,
            "status": "Port scan functionality will be implemented in Phase 3"
        }
    
    def service_detection(self, target, port):
        """Detect service running on a port.
        Stub method for Phase 3.
        
        Args:
            target (str): Target to scan
            port (int): Port to scan
            
        Returns:
            dict: Results of the service detection
        """
        # This is a stub that will be implemented in Phase 3
        logging.info(f"Service detection requested for {target}:{port}")
        
        # Save to database
        db.log_activity("network", "service_detection", {
            "target": target,
            "port": port
        })
        
        # Return placeholder results
        return {
            "target": target,
            "port": port,
            "status": "Service detection functionality will be implemented in Phase 3"
        }
    
    def check_host(self, host):
        """Check if a host is online.
        Simple implementation for MVP.
        
        Args:
            host (str): Host to check
            
        Returns:
            bool: True if host is online, False otherwise
        """
        try:
            # Try to resolve hostname
            socket.gethostbyname(host)
            return True
        except socket.error:
            return False

# Global network scanner instance
network_scanner = NetworkScanner() 